Rename package stream to package elasticapm
