+++ input/elasticapm/result_test.go
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package stream
+package elasticapm
 
 import (
 	"testing"
